+++ valid_s.py
@@ -27,7 +27,7 @@ def valid_s(end_epoch,epoch_num,model,dataloader,device,ckpt,num_class,writer,lo
         for batch_idx, (inputs, targets) in enumerate(loader):
 
             inputs, targets = inputs.to(device), targets.to(device)
-
+            inputs = inputs.float()
             targets = targets.float()
             outputs = model(inputs)
